limiting number of characters fixed
